Revert "Handle NativeScript as ScriptRef"
