+++ common-spec/src/main/java/com/bloxbean/cardano/client/spec/Script.java
@@ -1,6 +1,12 @@
 package com.bloxbean.cardano.client.spec;
 
+import co.nstant.in.cbor.CborException;
+import co.nstant.in.cbor.model.Array;
+import co.nstant.in.cbor.model.ByteString;
 import co.nstant.in.cbor.model.DataItem;
+import co.nstant.in.cbor.model.UnsignedInteger;
+import com.bloxbean.cardano.client.common.cbor.CborSerializationUtil;
+import com.bloxbean.cardano.client.exception.CborRuntimeException;
 import com.bloxbean.cardano.client.exception.CborSerializationException;
 import com.bloxbean.cardano.client.util.HexUtil;
 import com.fasterxml.jackson.annotation.JsonIgnore;
@@ -27,6 +33,26 @@ public interface Script {
         return finalBytes;
     }
 
+    /**
+     * Get serialized bytes for script reference. This is used in TransactionOutput's script_ref
+     * @return
+     * @throws CborSerializationException
+     */
+    default byte[] scriptRefBytes() throws CborSerializationException {
+        int type = getScriptType();
+        byte[] serializedBytes = serializeScriptBody();
+
+        Array array = new Array();
+        array.add(new UnsignedInteger(type));
+        array.add(new ByteString(serializedBytes));
+
+        try {
+            return CborSerializationUtil.serialize(array);
+        } catch (CborException e) {
+            throw new CborRuntimeException(e);
+        }
+    }
+
     @JsonIgnore
     default byte[] getScriptHash() throws CborSerializationException {
         return blake2bHash224(serialize());
@@ -36,13 +62,6 @@ public interface Script {
     default String getPolicyId() throws CborSerializationException {
         return HexUtil.encodeHexString(getScriptHash());
     }
-
-    /**
-     * Get serialized bytes for script reference. This is used in TransactionOutput's script_ref
-     * @return byte[]
-     * @throws CborSerializationException
-     */
-    byte[] scriptRefBytes() throws CborSerializationException;
 
     DataItem serializeAsDataItem() throws CborSerializationException;
     byte[] serializeScriptBody() throws CborSerializationException;
