+++ plutus/src/main/java/com/bloxbean/cardano/client/plutus/spec/PlutusScript.java
@@ -2,12 +2,8 @@ package com.bloxbean.cardano.client.plutus.spec;
 
 import co.nstant.in.cbor.CborDecoder;
 import co.nstant.in.cbor.CborException;
-import co.nstant.in.cbor.model.Array;
 import co.nstant.in.cbor.model.ByteString;
 import co.nstant.in.cbor.model.DataItem;
-import co.nstant.in.cbor.model.UnsignedInteger;
-import com.bloxbean.cardano.client.common.cbor.CborSerializationUtil;
-import com.bloxbean.cardano.client.exception.CborRuntimeException;
 import com.bloxbean.cardano.client.exception.CborSerializationException;
 import com.bloxbean.cardano.client.spec.Script;
 import com.bloxbean.cardano.client.util.HexUtil;
@@ -51,45 +47,6 @@ public abstract class PlutusScript implements Script {
         return serializeAsDataItem().getBytes();
     }
 
-    @Override
-    public byte[] scriptRefBytes() throws CborSerializationException {
-        int type = getScriptType();
-        byte[] serializedBytes = serializeScriptBody();
-
-        Array array = new Array();
-        array.add(new UnsignedInteger(type));
-        array.add(new ByteString(serializedBytes));
-
-        try {
-            return CborSerializationUtil.serialize(array);
-        } catch (CborException e) {
-            throw new CborRuntimeException(e);
-        }
-    }
-
-    public static PlutusScript deserializeScriptRef(byte[] scriptRefBytes) {
-        Array plutusScriptArray = (Array) CborSerializationUtil.deserialize(scriptRefBytes);
-        List<DataItem> dataItemList = plutusScriptArray.getDataItems();
-        if (dataItemList == null || dataItemList.size() == 0) {
-            throw new CborRuntimeException("PlutusScript deserialization failed. Invalid no of DataItem");
-        }
-
-        int type = ((UnsignedInteger) dataItemList.get(0)).getValue().intValue();
-        ByteString scriptBytes = ((ByteString) dataItemList.get(1));
-        try {
-            if (type == 1) {
-                return PlutusV1Script.deserialize(scriptBytes);
-            } else if (type == 2) {
-                return PlutusV2Script.deserialize(scriptBytes);
-            } else if (type == 3) {
-                return PlutusV3Script.deserialize(scriptBytes);
-            } else {
-                throw new CborRuntimeException("Invalid type : " + type);
-            }
-        } catch (Exception e) {
-            throw new CborRuntimeException("PlutusScript deserialization failed.", e);
-        }
-    }
-
     public abstract Language getLanguage();
+
 }
