+++ quicktx/src/main/java/com/bloxbean/cardano/client/quicktx/AbstractTx.java
@@ -291,15 +291,6 @@ public abstract class AbstractTx<T> {
         return (T) this;
     }
 
-    /**
-     * Checks if the transaction has any multi-asset minting or burning.
-     *
-     * @return true if there are multi-assets to be minted; false otherwise
-     */
-    boolean hasMultiAssetMinting() {
-        return multiAssets != null && !multiAssets.isEmpty();
-    }
-
     TxBuilder complete() {
         TxOutputBuilder txOutputBuilder = null;
