+++ quicktx/src/main/java/com/bloxbean/cardano/client/quicktx/QuickTxBuilder.java
@@ -228,7 +228,6 @@ public class QuickTxBuilder {
             TxBuilder txBuilder = (context, txn) -> {
             };
             boolean containsScriptTx = false;
-            boolean hasMultiAssetMint = false;
 
             Set<String> fromAddresses = new HashSet<>();
             for (AbstractTx tx : txList) {
@@ -254,8 +253,6 @@ public class QuickTxBuilder {
 
                 if (tx instanceof ScriptTx)
                     containsScriptTx = true;
-
-                hasMultiAssetMint = hasMultiAssetMint || tx.hasMultiAssetMinting();
             }
 
             int totalSigners = getTotalSigners();
@@ -361,7 +358,7 @@ public class QuickTxBuilder {
             //Balance outputs
             txBuilder = txBuilder.andThen(ScriptBalanceTxProviders.balanceTx(feePayer, totalSigners, containsScriptTx));
 
-            if ((containsScriptTx || hasMultiAssetMint) && removeDuplicateScriptWitnesses) {
+            if (containsScriptTx && removeDuplicateScriptWitnesses) {
                 txBuilder = txBuilder.andThen(DuplicateScriptWitnessChecker.removeDuplicateScriptWitnesses());
             }
